io.sphere.sdk.client.Tokens as interface #1053
