+++ commercetools-java-client-core/src/main/java/io/sphere/sdk/client/Tokens.java
@@ -1,58 +1,25 @@
 package io.sphere.sdk.client;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.type.TypeReference;
-import io.sphere.sdk.models.Base;
-import io.sphere.sdk.models.SphereException;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import javax.annotation.Nullable;
 import java.time.ZonedDateTime;
-import java.util.Optional;
-
-import static org.apache.commons.lang3.StringUtils.isEmpty;
 
 /** OAuth tokens returned by the authorization server. */
-public final class Tokens extends Base {
-    @JsonProperty("access_token")
-    private final String accessToken;
-    @JsonProperty("refresh_token")
-    private final String refreshToken;
-    @JsonProperty("expires_in")
-    @Nullable
-    private final Long expiresIn;
-    @JsonIgnore
-    @Nullable
-    private final ZonedDateTime expiresInZonedDateTime;
-
-    @JsonCreator
-    private Tokens(String accessToken, String refreshToken, @Nullable Long expiresIn) {
-        if (isEmpty(accessToken))
-            throw new SphereException("OAuth response must contain an access_token. Was empty.");
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-        this.expiresIn = expiresIn;
-        expiresInZonedDateTime = Optional.ofNullable(expiresIn).map(seconds -> ZonedDateTime.now().plusSeconds(seconds)).orElse(null);
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public String getRefreshToken() {
-        return refreshToken;
-    }
+@JsonDeserialize(as = TokensImpl.class)
+public interface Tokens {
+    String getAccessToken();
 
     @Nullable
-    public Long getExpiresIn() {
-        return expiresIn;
-    }
+    String getRefreshToken();
 
     @Nullable
-    ZonedDateTime getExpiresZonedDateTime() {
-        return expiresInZonedDateTime;
-    }
+    Long getExpiresIn();
+
+    @Nullable
+    public ZonedDateTime getExpiresZonedDateTime();
 
     /**
      * Creates a container which contains the full Java type information to deserialize this class from JSON.
@@ -64,7 +31,7 @@ public final class Tokens extends Base {
      *
      * @return type reference
      */
-    public static TypeReference<Tokens> typeReference() {
+    static TypeReference<Tokens> typeReference() {
         return new TypeReference<Tokens>() {
             @Override
             public String toString() {
@@ -74,7 +41,7 @@ public final class Tokens extends Base {
     }
 
     @JsonIgnore
-    public static Tokens of(String accessToken, String refreshToken, Long expiresIn) {
-        return new Tokens(accessToken, refreshToken, expiresIn);
+    static Tokens of(String accessToken, String refreshToken, Long expiresIn) {
+        return new TokensImpl(accessToken, refreshToken, expiresIn);
     }
 }
